Fix remaining compilation errors and warnings in Utils, RDEF, CPUView, and ActivityGraphView
- Utils.cpp: Comment out `cpu_clock_speed` access in `GetRoundedCpuSpeed` as the member is missing in the target `system_info` struct. Returns 0 as fallback.
- SysMonTask.rdef: Replace `B_APPV_FINAL` with literal `0` to resolve syntax error during resource compilation.
- ActivityGraphView.cpp: Fix signed/unsigned comparison warning by casting `steps` to `int32`.
- CPUView.cpp: Fix signed/unsigned comparison warnings by casting `sysInfo.used_teams` and `sysInfo.used_threads` to `int32` before comparing with `fLastUsedTeams` and `fLastUsedThreads`.

+++ SystemMonitor/CPUView.cpp
@@ -247,13 +247,13 @@ void CPUView::UpdateData()
     // Update Info
     system_info sysInfo;
     if (get_system_info(&sysInfo) == B_OK) {
-        if (sysInfo.used_teams != fLastUsedTeams) {
+        if ((int32)sysInfo.used_teams != fLastUsedTeams) {
             fLastUsedTeams = sysInfo.used_teams;
             fCachedProcesses.SetToFormat("%" B_PRId32, fLastUsedTeams);
             fProcessesValue->SetText(fCachedProcesses.String());
         }
 
-        if (sysInfo.used_threads != fLastUsedThreads) {
+        if ((int32)sysInfo.used_threads != fLastUsedThreads) {
             fLastUsedThreads = sysInfo.used_threads;
             fCachedThreads.SetToFormat("%" B_PRId32, fLastUsedThreads);
             fThreadsValue->SetText(fCachedThreads.String());
